JS structure improvements
