+++ src/js/controllers/data/episode/index.js
@@ -1,5 +1,5 @@
 import { Controller } from '@hotwired/stimulus';
-import Storage from '../../../storage';
+import Storage from '../../../models/Storage';
 
 const EpisodesController = class extends Controller {
   static targets = ['tags', 'items', 'count'];
