+++ src/js/controllers/data/post/index.js
@@ -1,5 +1,5 @@
 import { Controller } from '@hotwired/stimulus';
-import Storage from '../../../storage';
+import Storage from '../../../models/Storage';
 
 const PostsController = class extends Controller {
   static targets = ['title', 'content', 'published', 'items', 'count'];
