+++ src/js/controllers/navigation.js
@@ -1,4 +1,5 @@
 import { Controller } from '@hotwired/stimulus';
+import Route from '../models/Route';
 
 const NavigationController = class extends Controller {
   static values = {
@@ -7,30 +8,8 @@ const NavigationController = class extends Controller {
 
   navigate(e) {
     e.preventDefault();
-    const route = this.element.href.split(`${window.location.host}/`)
-    const parts = route[1].split('/');
-    this.get(parts[0], parts[1]);
-  }
-
-  get(page, id) {
-    const route = id ? `/${page}/${id}` : `/${page}`;
-    
-    const xhr = new XMLHttpRequest();
-    xhr.onreadystatechange = function() {
-      if (this.readyState == 4 && this.status == 200) {
-  
-        document.getElementById("spa-content-container").innerHTML = xhr.responseText;
-        const title = `Pattern Radio - Online Music Broadcasting`;
-        window.history.pushState(
-          { 'content': xhr.responseText, 'title': title },
-          title,
-          route
-        );
-      }
-    };
-    
-    xhr.open('GET', `/page${route}`, true);
-    xhr.send();
+    const route = new Route(this.element.href.split(`${window.location.host}/`)[1]);
+    route.content();
   }
 };
