+++ src/js/models/Registry.js
